requirements.txt
docx2txt==0.8
PyPDF2
pypdf==3.17.4
python-dotenv
faiss-cpu
openai==1.12.0
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.23
langchain_experimental
langchain-openai==0.0.6
openapi-schema-pydantic==1.2.4
pandas==2.1.3
requests==2.31.0
streamlit==1.28.2
tenacity==8.2.3
tiktoken
azure-identity==1.15.0
azure-search-documents==11.4.0b8
streamlit-agraph==0.0.45
gremlinpython==3.7.1
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
